Move ReviewPersistentLayer and mount arrival notifications
Replace auto-navigation on review push with the notification layer.
